refactor(PAT-62): Aplly SOLID and BaseTicket is added.
